+++ app/services/establishments_fetcher.rb
@@ -3,7 +3,7 @@ class EstablishmentsFetcher
     return unless projects&.any?
 
     urns = projects.map(&:urn)
-    establishments = AcademiesApi::Client.new.get_establishments(urns).object
+    establishments = Api::AcademiesApi::Client.new.get_establishments(urns).object
 
     projects.each do |project|
       project.establishment = establishments.find { |establishment| establishment.urn == project.urn.to_s }